+++ frontend/src/features/agent/libs/AgentUtils.ts
@@ -1,4 +1,4 @@
-const logisticCurve = (x: number) => {
+export const logisticCurve = (x: number) => {
   const L = 1; // 最大値
   const k = 0.8; // 勾配を決定する定数
   const x0 = 3; // 中点
