+++ frontend/src/features/agent/types/index.ts
@@ -1,0 +1,12 @@
+export type AgentInput = {
+  tools: string[];
+};
+
+export type AgentTool = {
+  name: string;
+  description: string;
+};
+
+export type Agent = {
+  tools: AgentTool[];
+};
