+++ frontend/src/features/agent/xstates/agentThinkProgress.ts
