+++ frontend/src/hooks/usePostMessageStreaming.ts
@@ -2,7 +2,7 @@ import { Auth } from 'aws-amplify';
 import { PostMessageRequest } from '../@types/conversation';
 import { create } from 'zustand';
 import i18next from 'i18next';
-import { AgentThinkingEventKeys } from '../xstates/agentThinkProgress';
+import { AgentThinkingEventKeys } from '../features/agent/xstates/agentThinkProgress';
 
 const WS_ENDPOINT: string = import.meta.env.VITE_APP_WS_ENDPOINT;
 const CHUNK_SIZE = 32 * 1024; //32KB
