+++ frontend/src/pages/ChatPage.tsx
@@ -27,7 +27,7 @@ import StatusSyncBot from '../components/StatusSyncBot';
 import Alert from '../components/Alert';
 import useBotSummary from '../hooks/useBotSummary';
 import useModel from '../hooks/useModel';
-import { progress } from '../utils/AgentUtils';
+import { progress } from '../features/agent/libs/AgentUtils';
 
 const MISTRAL_ENABLED: boolean =
   import.meta.env.VITE_APP_ENABLE_MISTRAL === 'true';
